Dealt with warnings, simplified click system [bug]
There is still a visual bug with the zooming in and everything except
the map not zooming correctly

+++ MapDrawer.h
@@ -36,7 +36,7 @@ public:
 		inline bool GetRedrawState(){ return needsRedraw; }
 	#pragma endregion
 
-	inline void SetZoom(int zoomToSet){ mapSprite.setScale(zoomToSet, zoomToSet); currentZoom = zoomToSet; }
+		inline void SetZoom(int zoomToSet){ mapSprite.setScale((float)zoomToSet, (float)zoomToSet); currentZoom = zoomToSet; }
 	inline int GetZoomValue(){ return currentZoom; };
 
 	sf::Sprite mapSprite;
